+++ cs246-final-project/cc3k/Game/GameObject/Player/Drow.cc
@@ -9,7 +9,7 @@ bool Drow::attackEnemy(Enemy *e) {
 }
 
 void Drow::attackedBy(Enemy *e) {
-    hp_now -= ceil((100 / (100 + getDef())) * e->getAtk());
+    hp_now -= static_cast<int>(ceil((100.0 / (100.0 + getDef())) * e->getAtk()));
 }
 
 int Drow::getTileID() { return id; };
